+++ src/main/java/RacingCar.java
@@ -1,34 +1,34 @@
-import java.util.Random;
-
-public class RacingCar {
-    private String name;
-    private int distance;
-
-    public RacingCar(String name) {
-        this.name = name;
-        this.distance = 0;
-    }
-
-    public int randomNum() {
-        Random random = new Random();
-        int num = random.nextInt(10);
-
-        return num;
-    }
-
-    public void move(int num) {
-        if(num >= 4) {
-            this.distance++;
-        }
-    }
-
-    public String getName() {
-        return this.name;
-    }
-    public int getDistance() {
-        return this.distance;
-    }
-
-}
-
-
+//import java.util.Random;
+//
+//public class RacingCar {
+//    private String name;
+//    private int distance;
+//
+//    public RacingCar(String name) {
+//        this.name = name;
+//        this.distance = 0;
+//    }
+//
+//    public int randomNum() {
+//        Random random = new Random();
+//        int num = random.nextInt(10);
+//
+//        return num;
+//    }
+//
+//    public void move(int num) {
+//        if(num >= 4) {
+//            this.distance++;
+//        }
+//    }
+//
+//    public String getName() {
+//        return this.name;
+//    }
+//    public int getDistance() {
+//        return this.distance;
+//    }
+//
+//}
+//
+//
